rename webmention-for-url files
use `-` instead of `_`
