+++ lib/eleventy.config.js
@@ -12,7 +12,7 @@ module.exports = function (eleventyConfig) {
     // Filters
     eleventyConfig.addNunjucksFilter("date", dateFilter);
     eleventyConfig.addFilter("hostname", require('./filters/hostname.js'));
-    eleventyConfig.addFilter("webmentions_for_url", require('./filters/webmentions_for_url.js'));
+    eleventyConfig.addFilter("webmentions_for_url", require('./filters/webmentions-for-url.js'));
 
     // Plugins
     eleventyConfig.addPlugin(require('@11ty/eleventy-plugin-syntaxhighlight'));
